Fixed problem in handling of downloaded datasets.

+++ workflow/rules/taxa.smk
@@ -40,8 +40,7 @@ rule col_get_data:
     data_dir = config['col_data_dir'],
     dl_dir = f"{config['col_data_dir']}/.."
   log:
-    out = '../log/col_get_data.out',
-    err = '../log/col_get_data.err'
+    '../log/col_get_data.log'
   benchmark:
     '../log/col_get_data.prf'
   shell:
@@ -54,25 +53,27 @@ rule col_get_data:
       --out-dir {params.dl_dir}
 
     # check for sentinel file
-    for f_new in $(ls {params.dl_dir}/*.new); do
-      echo "Found sentinel file: $f_new"
-      f_data=${{f_new%.new}}
+    if [[ -f "{params.dl_dir}/*.new" ]]; then
+      for f_new in $(ls {params.dl_dir}/*.new || true); do
+        echo "Found sentinel file: $f_new"
+        f_data=${{f_new%.new}}
 
-      # clear or create data directory
-      if [[ -d {params.data_dir} ]]; then
-        rm -rf {params.data_dir}/*
-      else
-        mkdir {params.data_dir}
-      fi
+        # clear or create data directory
+        if [[ -d {params.data_dir} ]]; then
+          rm -rf {params.data_dir}/*
+        else
+          mkdir {params.data_dir}
+        fi
 
-      # unzip data files
-      unzip $f_data -d {params.data_dir}
+        # unzip data files
+        unzip $f_data -d {params.data_dir}
 
-      # clean up
-      rm $f_new $f_data
-      break
-    done
-    ) >{log.out} 2>{log.err}
+        # clean up
+        rm $f_new $f_data
+        break
+      done
+    fi
+    ) >{log} 2>&1
     """
 
 # Download latest Encyclopedia of Life (EOL) data files
@@ -84,8 +85,7 @@ rule eol_get_data:
   params:
     data_dir = config['eol_data_dir']
   log:
-    out = '../log/eol_get_data.out',
-    err = '../log/eol_get_data.err'
+    '../log/eol_get_data.log'
   benchmark:
     '../log/eol_get_data.prf'
   shell:
@@ -105,7 +105,7 @@ rule eol_get_data:
 
     # unzip data files
     unzip {params.data_dir}/../traits_all.zip -d {params.data_dir}/..
-    ) >{log.out} 2>{log.err}
+    ) >{log} 2>&1
     """
 
 # Parse Catalogue of Life (CoL) taxa file and extract descendents of Arthropoda
@@ -115,8 +115,7 @@ rule col_taxa_extract_arthropods:
   output:
     tax_art = config['col_taxa_arthropoda']
   log:
-    out = '../log/col_taxa_extract_arthropods.out',
-    err = '../log/col_taxa_extract_arthropods.err'
+    '../log/col_taxa_extract_arthropods.log'
   benchmark:
     '../log/col_taxa_extract_arthropods.prf'
   shell:
@@ -126,7 +125,7 @@ rule col_taxa_extract_arthropods:
       python scripts/col_extract_arthropods.py \
         --col-file {input.col} \
         --out-file {output.tax_art}
-    ) >{log.out} 2>{log.err}
+    ) >{log} 2>&1
     """
 
 # Parse Encyclopedia of Life (EOL) taxa, traits and relationships and extract descendents of Arthropoda
